### 2019-07-25 #1597 @whedon
[PRE REVIEW]: AltamISA: a Python API for ISA-Tab files
**Submitting author:** @holtgrewe (<a href="http://orcid.org/0000-0002-3051-1763">Manuel Holtgrewe</a>)
**Repository:** <a href="https://github.com/bihealth/altamisa" target ="_blank">https://github.com/bihealth/altamisa</a>
**Version:** v0.2.0
