### 2019-07-26 #1597 @holtgrewe
@dmnfarrel @tleonardi and @dsurujon would be three suitable candidates recently added to the list of volunteers that have specified both Python and bioinformatics as areas of expertise.
