### 2019-07-26 #1597 @tleonardi
This seems a very interesting tool, I'd be happy to review the submission.
