### 2019-07-26 #1597 @dmnfarrell
Yes I'd be willing to review this submission.
