### 2019-07-26 #1597 @kyleniemeyer
Thanks @tleonardi @dmnfarrell and @dsurujon for responding so quickly! I am still waiting to get an editor assigned to manage the submission, but I will definitely take you all up on the offer to review.
