### 2019-07-29 #1597 @holtgrewe
@labarba 👋 Thanks for raising this point. Would the following be sufficient and appropriately terse?

"AltamISA is a Python library for reading, validating, representing, and writing the ISA-Tab file format. ISA-Tab is an open, TSV-based file format for modeling and sharing life science experiments that is in use by various life science databases such as MetaboLights."
