### 2019-07-29 #1597 @labarba
That's a start. How can a file format be used for _modeling_, though?
