### 2019-07-30 458ab88
Adjusting introduction to a broader audience
cf. https://github.com/openjournals/joss-reviews/issues/1597
