### 2019-07-30 #1597 @labarba
That's much better!

I want to say that "shortly" is used in English to mean "in a little while" or "soon," so I think your word choice is not working at the start of paragraph 2.
