### 2019-07-30 #1597 @labarba
👋 @lpantano — I believe this submission is within your domains of knowledge. Could you handle as editor?

It is a Python library for reading, validating, representing, and writing the ISA-Tab file format, used in life science experiments.
