### 2019-08-05 #1610 @tleonardi
Hi @holtgrewe, I'm in the process of reviewing the submission and still have to test its functionality.
In the meantime I have a couple of comments.

I've noticed that the paper lists "Mathias Kuhring, Mikko Nieminen, Jennifer Kirwan, Dieter Beule and Manuel Holtgrewe" as authors. However, `AUTHORS.rst` and the Credits section of the docs only report Kuhring, Nieminen and Holtgrewe (with commits only by Kuhring and Holtgrewe). Can you confirm that all co-authors have actively contributed to AlamISA development in line with JOSS [authorship guidelines](https://joss.readthedocs.io/en/latest/submitting.html#authorship)?

On a separate note, can you add a DOI to the references in the bibliography?
