### 2019-08-05 e592e38
Properly adjusting authorship.
