+++ AUTHORS.rst
@@ -2,6 +2,8 @@
 Credits
 =======
 
+- Dieter Beule
 - Manuel Holtgrewe
+- Jennifer Kirwan
 - Mathias Kuhring
-- Mikko Niminen
+- Mikko Nieminen
