### 2019-08-05 bihealth/altamisa#58 @holtgrewe
Hi, thanks for the suggestions. I'll address this with a `v0.2.1` release.

Thanks for suggesting poetry. My impression was always that it is not ready yet and I was happy with twine+versioneer so far. I'll put this on my list for the future, though.
