### 2019-08-05 bihealth/altamisa#58 @holtgrewe
Fixed in `v0.2.1`, also on PyPI.
