### 2019-08-05 #1610 @holtgrewe
@tleonardi Thanks for finding this. Yes, while Beule and Kirwan have contributed no lines of code, they have contributed a lot in terms of software design and domain-specific knowledge, without this, the code would not be where it is now. Also, Nieminen is sitting next to Kuhring in the office and has worked (accumulated certainly hours) with Kuhring in pair programming, discussion, and consulting.

I thus see all authors on the paper fit for inclusion as authors and all have contributed to the content of the package and have made a contribution far beyond just supervision and funding.
