+++ paper.bib
@@ -6,7 +6,8 @@
    Volume="44",
    Number="2",
    Pages="121--126",
-   Month="Jan"
+   Month="Jan",
+   doi="10.1038/ng.1054"
 }
 
 @Article{Wilkinson:2016,
@@ -16,7 +17,8 @@
    Year="2016",
    Volume="3",
    Pages="160018",
-   Month="Mar"
+   Month="Mar",
+   doi="10.1038/sdata.2016.18"
 }
 
 @unpublished{Nieminen:2020,
@@ -33,7 +35,8 @@
    Volume="41",
    Number="Database issue",
    Pages="D781--786",
-   Month="Jan"
+   Month="Jan",
+   doi="10.1002/0471250953.bi1413s53"
 }
 
 @ARTICLE{Gansner:1993,
@@ -41,7 +44,8 @@
     title = {A Technique for Drawing Directed Graphs},
     journal = {IEEE TRANSACTIONS ON SOFTWARE ENGINEERING},
     year = {1993},
-    pages = {214--230}
+    pages = {214--230},
+    doi = "10.1109/32.221135"
 }
 
 @Article{Gonzalez:2014,
@@ -50,5 +54,6 @@
    Journal="BMC Bioinformatics",
    Year="2014",
    Volume="15 Suppl 14",
-   Pages="S4"
+   Pages="S4",
+   doi="10.1186/1471-2105-15-S14-S4"
 }
