### 2019-08-05 #1610 @holtgrewe
@tleonardi And of course, I can add DOIs (just added to `master`).
