### 2019-08-06 bihealth/altamisa#60 @tleonardi
import statements in the docs
The Getting Started page in the documentation says to import the isatab module with:
```
from altamisa import isatab
```
However, the rest of the examples call isatab functions like this:
```
with open("path/to/an/investigation.txt", "rt") as investigation_file:
        investigation = InvestigationReader.from_stream(investigation_file).read()

InvestigationValidator(investigation).validate()
```
To make things easier for a beginner you could either prefix `isatab.` to the function calls or import the module as:
```
 from altamisa.isatab import *
```

I spotted this issue as part of the JOSS review process (openjournals/joss-reviews#1610).
