### 2019-08-06 #1610 @tleonardi
@holtgrewe thanks for the clarification about authorship, I think this is perfectly in line with the journal's policy.
