### 2019-08-06 #1610 @tleonardi
@lpantano could you advise on JOSS policy about citing manuscripts _in preparation_? The authors are citing their _SODAR_ manuscript as _Nieminen, Holtgrewe and Beule, 2020_. Is this ok?
