### 2019-08-06 988fd11
Various text editing suggestions
