### 2019-08-06 #1610 @lpantano
Hi @tleonardi  and @holtgrewe, the board of editors recommends to deposit that paper into biorxiv if you really want to cite it. If that is not possible, sadly it needs to be removed from the reference and mentioned in the paper that it is under preparation or working on that. I hope that works.
