### 2019-08-06 #1610 @holtgrewe
@lpantano Sounds very well in line with the spirit of JOSS. We'll be happy to remove it as we don't have much to show for it right now.
