### 2019-08-06 5790bac
Removing citation of unpublished material.
cf. openjournals/joss-reviews#1610
