+++ paper.bib
@@ -21,12 +21,6 @@
    doi="10.1038/sdata.2016.18"
 }
 
-@unpublished{Nieminen:2020,
-    Author = {{Nieminen}, M. and {Holtgrewe}, M. and {Beule}, D.},
-    Title = {SODAR: Management of mass and meta data for Omics data},
-    Year = 2020,
-    notes = {in preparation}}
-
 @Article{Haug:2013,
    Author="Haug, K.  and Salek, R. M.  and Conesa, P.  and Hastings, J.  and de Matos, P.  and Rijnbeek, M.  and Mahendraker, T.  and Williams, M.  and Neumann, S.  and Rocca-Serra, P.  and Maguire, E.  and Gonzalez-Beltran, A.  and Sansone, S. A.  and Griffin, J. L.  and Steinbeck, C. ",
    Title="{{M}etabo{L}ights--an open-access general-purpose repository for metabolomics studies and associated meta-data}",
