### 2019-08-06 314e96e
Ignoring errors in coverage computations.
This is important for pull request builds.
