+++ tox.ini
@@ -14,4 +14,4 @@ commands =
     pytest
     py36: flake8 setup.py altamisa tests
     py36: black -l 100 --check .
-    py36: python-codacy-coverage -r coverage.xml
+    py36: - python-codacy-coverage -r coverage.xml
