### 2019-08-07 #1610 @tleonardi
@lpantano I have now completed my review of this submission. AltamISA is a useful research software that facilitates the systematic annotation of experimental procedure in line with the FAIR (Findable, Accessible, Interoperable, and Reusable) guidelines. The paper is clear and well written - as is the code, which is also well documented and thoroughly tested.
In my review I have only found some minor issues ([#58](https://github.com/bihealth/altamisa/issues/58), [#59](https://github.com/bihealth/altamisa/issues/59), [#60](https://github.com/bihealth/altamisa/issues/60), [#61](https://github.com/bihealth/altamisa/issues/61)) which the authors have now fixed/addressed.
I'm happy to reccommend the paper for publication in JOSS.
