### 2019-08-11 bihealth/altamisa#64 @dsurujon
[JOSS Review] Documentation
* AltamISA version: 0.2.1
* Python version: 3.6
* Operating System: Windows 10

I have found 3 minor issues in the documentation:
* In the documentation, the examples provided are very easy to follow. The authors have also kindly provided some example files to use in the repository. It would be nice to include a reference to a real example file instead of having the placeholder ```path/to/an/investigation.txt``` in the examples.

* In the second example code snippet (parsing study and assay data), there is a typo in the open statement - there is an extra closing parenthesis: ```with open("path/to/a/study_file.txt"), "rt")``` should be ```with open("path/to/a/study_file.txt", "rt")``` (same applies to the assay parsing in the same example)

* Under the "Manual ISA Model creation" section, the authors have supplied a handy example function called ```create_and_write```. To use this example successfully, the user needs to make sure the "out_path" specifies an existing directory. I would suggest the authors either specify "out_path" needs to exist, or include an internal check, that generates the directory in the case it doesn't. Secondly, in the same example, there are 3 instances (each time a file is written) that the function ```join``` is used. I suggest to change these to ```os.path.join```

This is a part of the JOSS review for this project (openjournals/joss-reviews#1610).
