### 2019-08-11 #1610 @dsurujon
I have gone through the manuscript, documentation and examples for AltamISA. I think this is a very nice and convenient package, and I'd like to thank the authors for introducing me to the ISA framework - I will try to incorporate ISA and AltamISA into my work in the future!
I spotted a couple of typos in the documentation, and have one suggestion. @holtgrewe could you please address those issues I listed in this issue (bihealth/altamisa/issues/64)

Moreover, I have 2 small issues regarding the manuscript:
* Under the section "License and Availability" it says the current version is "0.2.0", however after addressing @tleonardi 's comments the version number is now 0.2.1
* Towards the end of the "Aims and Features" section, the authors note that they excluded JSON and RFD file formats from this project because they are not widely adapted. I would argue that at least for JSON, it is a very well-accepted and ubiquitously used file format. It is possible taht the authors mean specifically within applications that use the ISA framework, these formats are more rare than the ISA-Tab format. If this is the case, I would suggest rephrasing this statement to clarify this point.

Once these issues are addressed I would be more than happy to recommend this paper for publication. Thank you @holtgrewe and colleagues for generating this nice package!
