### 2019-08-12 #1610 @dmnfarrell
I have reviewed this software and find the paper clear and well written. The software installs via pip with no issues. I have no experience with the ISA-tab format but the code examples I tested worked and the API documentation look well prepared. Some of the issues I noticed have already been addressed by the other reviewers. The only other comments I have are as follows:
* it is not clear how the tests are to be run from the documentation, if this is possible
* issue bihealth/altamisa/issues/60 is yet to be addressed in terms of the import statements in the docs

Otherwise I am happy to recommend for publication.
