### 2019-08-14 80fba7f
Fix writing of process performer and date
