### 2019-08-14 7803239
Fix minor documentation issues (#60, #64)
