+++ docs/examples.rst
@@ -4,6 +4,18 @@
 Examples
 ========
 
+To run the examples, download a full
+`test dataset <https://github.com/bihealth/altamisa/tree/master/tests/data/BII-I-1>`_
+(all files) from the AltamISA repository to your working directory.
+
+Import AltamISA (if not already done) and other required modules.
+
+.. code-block:: python
+
+    from altamisa.isatab import *
+    import os
+
+
 Parsing and validation
 ----------------------
 
@@ -13,39 +25,39 @@ other. Reading and validating a single investigation file is performed as follow
 
 .. code-block:: python
 
-	with open("path/to/an/investigation.txt", "rt") as investigation_file:
+	with open("i_investigation.txt", "rt") as investigation_file:
 		investigation = InvestigationReader.from_stream(investigation_file).read()
-	
+
 	InvestigationValidator(investigation).validate()
-	
+
 For study and assay parsing, unique ids need to be set to enable unambiguous identification of study
 , assay and, in particular, their nodes in later applications (such as a complete graph creation):
 
 .. code-block:: python
-	
-	with open("path/to/a/study_file.txt"), "rt") as inputf:
+
+	with open("s_BII-S-1.txt", "rt") as inputf:
 		study = StudyReader.from_stream("S1", inputf).read()
-				
-	with open("path/to/a/study_file.txt"), "rt") as inputf:
+
+	with open("a_transcriptome.txt", "rt") as inputf:
 		assay = AssayReader.from_stream("S1", "A1", inputf).read()
 
 
 However, in real use cases an ISA-Tab dataset contains related investigation, study and assay files
 and thus should be handled as unit. In particular, validation of studies and assays requires
-information from parent elements such as the investigation or study, respectivly. Thus, joint
+information from parent elements such as the investigation or study, respectively. Thus, joint
 parsing and validation of a complete ISA-Tab dataset may look like this:
 
 .. code-block:: python
 
     # Read investigation
-    with open("path/to/an/investigation.txt", "rt") as investigation_file:
+    with open("i_investigation.txt", "rt") as investigation_file:
         investigation = InvestigationReader.from_stream(investigation_file).read()
 
     # Validate investigation
     InvestigationValidator(investigation).validate()
 
     # Read studies and assays
-    path_in = os.path.normpath(os.path.dirname("path/to/an/investigation.txt"))
+    path_in = os.path.normpath(os.path.dirname("i_investigation.txt"))
     studies = {}
     assays = {}
     for s, study_info in enumerate(investigation.studies):
@@ -67,7 +79,7 @@ parsing and validation of a complete ISA-Tab dataset may look like this:
             StudyValidator(investigation, study_info, studies[s]).validate()
         for a, assay_info in enumerate(study_info.assays):
             if assay_info.path:
-                AssayValidator(investigation, study_info, assay_info, assays[s][a]).validate()	
+                AssayValidator(investigation, study_info, assay_info, assays[s][a]).validate()
 
 
 Writing
@@ -80,22 +92,24 @@ investigation, as files might be overwritten otherwise):
 .. code-block:: python
 
     # Write investigation
-    with open("path/to/output/investigation.txt", "wt", newline='') as output_investigation_file:
+    path_out = "/tmp/altamisa_example/"
+    os.makedirs(path_out, exist_ok=True)
+    with open(os.path.join(path_out, "i_investigation.txt"), "wt", newline='') as output_investigation_file:
         InvestigationWriter.from_stream(
-            investigation, output_investigation_file, quote=args.quotes
+            investigation, output_investigation_file
         ).write()
 
     # Write studies and assays
     for s, study_info in enumerate(investigation.studies):
         if study_info.info.path:
[…]
