+++ docs/getting_started.rst
@@ -8,6 +8,6 @@ After installation, you can use AltamISA in your project simply by importing the
 
 .. code-block:: python
 
-    from altamisa import isatab
+    from altamisa.isatab import *
 
 That's all, continue and look at the list of examples.
