### 2019-08-15 aa76a77
Fix minor documentation issues (#60, #64) (#65)
