+++ docs/examples.rst
@@ -4,6 +4,17 @@
 Examples
 ========
 
+To run the examples, download a full
+`test dataset <https://github.com/bihealth/altamisa/tree/master/tests/data/BII-I-1>`_
+(all files) from the AltamISA repository to your working directory.
+
+Import AltamISA (if not already done) and other required modules.
+
+.. literalinclude:: examples/process_isa_model.py
+    :language: python
+    :lines: 3-5
+
+
 Parsing and validation
 ----------------------
 
@@ -11,64 +22,25 @@ AltamISA provides separate functions to parse, validate and export ISA-tab inves
 assay files, respectivly. Investigation, study and assay files may be parsed independently from each
 other. Reading and validating a single investigation file is performed as follows:
 
-.. code-block:: python
+.. literalinclude:: examples/process_isa_model.py
+    :language: python
+    :lines: 8-12
 
-	with open("path/to/an/investigation.txt", "rt") as investigation_file:
-		investigation = InvestigationReader.from_stream(investigation_file).read()
-	
-	InvestigationValidator(investigation).validate()
-	
 For study and assay parsing, unique ids need to be set to enable unambiguous identification of study
 , assay and, in particular, their nodes in later applications (such as a complete graph creation):
 
-.. code-block:: python
-	
-	with open("path/to/a/study_file.txt"), "rt") as inputf:
-		study = StudyReader.from_stream("S1", inputf).read()
-				
-	with open("path/to/a/study_file.txt"), "rt") as inputf:
-		assay = AssayReader.from_stream("S1", "A1", inputf).read()
-
+.. literalinclude:: examples/process_isa_model.py
+    :language: python
+    :lines: 15-20
 
 However, in real use cases an ISA-Tab dataset contains related investigation, study and assay files
 and thus should be handled as unit. In particular, validation of studies and assays requires
-information from parent elements such as the investigation or study, respectivly. Thus, joint
+information from parent elements such as the investigation or study, respectively. Thus, joint
 parsing and validation of a complete ISA-Tab dataset may look like this:
 
-.. code-block:: python
-
-    # Read investigation
-    with open("path/to/an/investigation.txt", "rt") as investigation_file:
-        investigation = InvestigationReader.from_stream(investigation_file).read()
-
-    # Validate investigation
-    InvestigationValidator(investigation).validate()
-
-    # Read studies and assays
-    path_in = os.path.normpath(os.path.dirname("path/to/an/investigation.txt"))
-    studies = {}
-    assays = {}
-    for s, study_info in enumerate(investigation.studies):
-        if study_info.info.path:
-            with open(os.path.join(path_in, study_info.info.path), "rt") as inputf:
-                studies[s] = StudyReader.from_stream("S{}".format(s + 1), inputf).read()
-        if study_info.assays:
-            assays[s] = {}
-        for a, assay_info in enumerate(study_info.assays):
-            if assay_info.path:
-                with open(os.path.join(path_in, assay_info.path), "rt") as inputf:
-                    assays[s][a] = AssayReader.from_stream(
-                        "S{}".format(s + 1), "A{}".format(a + 1), inputf
-                    ).read()
-
-    # Validate studies and assays
-    for s, study_info in enumerate(investigation.studies):
-        if study_info.info.path:
-            StudyValidator(investigation, study_info, studies[s]).validate()
-        for a, assay_info in enumerate(study_info.assays):
-            if assay_info.path:
-                AssayValidator(investigation, study_info, assay_info, assays[s][a]).validate()	
-
+.. literalinclude:: examples/process_isa_model.py
+    :language: python
+    :lines: 23-53
 
 Writing
 -------
[…]
