### 2019-08-15 bihealth/altamisa#60 @mkuhring
Thanks you very much for pointing out this issue. We updated the documentation accordingly (aa76a775d0cd40c1a0858409f90f05b2e3080c9d).
