### 2019-08-15 bihealth/altamisa#64 @mkuhring
Thanks you very much for pointing out these issues. We updated the documentation accordingly (aa76a775d0cd40c1a0858409f90f05b2e3080c9d). Please feel to (re-)open an issue in case you see need for additional improvements.
