### 2019-08-15 #1610 @mkuhring
@lpantano, @tleonardi, @dsurujon, @dmnfarrell, @holtgrewe
Thanks to all reviewers for your constructive criticism. We updated our documentation with respect to your remaining issues: https://github.com/bihealth/altamisa/issues/60, https://github.com/bihealth/altamisa/issues/64. We will tag a new release soon. Feel free to open additional issues if you see need for additional improvements.
