### 2019-08-15 aa354de
Update version and DOI in paper
