### 2019-08-15 887a33b
Releasing v0.2.2.
