### 2019-08-15 #1610 @lpantano
Thank you, @mkuhring !

can you confirm the new version is v0.2.2? or let me know when you have the latest version that addressed the reviewers' comments.

Cheers
