### 2019-08-15 #1610 @mkuhring
Yes, v0.2.2 is the new version.

Lorena Pantano <notifications@github.com> schrieb am Do., 15. Aug. 2019,
20:51:

> Thank you, @mkuhring <https://github.com/mkuhring> !
>
> can you confirm the new version is v0.2.2? or let me know when you have
> the latest version that addressed the reviewers' comments.
>
> Cheers
>
