### 2019-08-19 #1610 @lpantano
Hey @openjournals/joss-eics this paper has been accepted. Thanks all!
