### 2019-08-20 1280d00
Update paper.bib
Fixed some reference metadata
