+++ paper.bib
@@ -1,53 +1,53 @@
 @Article{Sansone:2012,
-   Author="Sansone, S. A.  and Rocca-Serra, P.  and Field, D.  and Maguire, E.  and Taylor, C.  and Hofmann, O.  and Fang, H.  and Neumann, S.  and Tong, W.  and Amaral-Zettler, L.  and Begley, K.  and Booth, T.  and Bougueleret, L.  and Burns, G.  and Chapman, B.  and Clark, T.  and Coleman, L. A.  and Copeland, J.  and Das, S.  and de Daruvar, A.  and de Matos, P.  and Dix, I.  and Edmunds, S.  and Evelo, C. T.  and Forster, M. J.  and Gaudet, P.  and Gilbert, J.  and Goble, C.  and Griffin, J. L.  and Jacob, D.  and Kleinjans, J.  and Harland, L.  and Haug, K.  and Hermjakob, H.  and Ho Sui, S. J.  and Laederach, A.  and Liang, S.  and Marshall, S.  and McGrath, A.  and Merrill, E.  and Reilly, D.  and Roux, M.  and Shamu, C. E.  and Shang, C. A.  and Steinbeck, C.  and Trefethen, A.  and Williams-Jones, B.  and Wolstencroft, K.  and Xenarios, I.  and Hide, W. ",
-   Title="{{T}oward interoperable bioscience data}",
-   Journal="Nat. Genet.",
-   Year="2012",
-   Volume="44",
-   Number="2",
-   Pages="121--126",
-   Month="Jan",
-   doi="10.1038/ng.1054"
+   Author={Sansone, S. A.  and Rocca-Serra, P.  and Field, D.  and Maguire, E.  and Taylor, C.  and Hofmann, O.  and Fang, H.  and Neumann, S.  and Tong, W.  and Amaral-Zettler, L.  and Begley, K.  and Booth, T.  and Bougueleret, L.  and Burns, G.  and Chapman, B.  and Clark, T.  and Coleman, L. A.  and Copeland, J.  and Das, S.  and de Daruvar, A.  and de Matos, P.  and Dix, I.  and Edmunds, S.  and Evelo, C. T.  and Forster, M. J.  and Gaudet, P.  and Gilbert, J.  and Goble, C.  and Griffin, J. L.  and Jacob, D.  and Kleinjans, J.  and Harland, L.  and Haug, K.  and Hermjakob, H.  and Ho Sui, S. J.  and Laederach, A.  and Liang, S.  and Marshall, S.  and McGrath, A.  and Merrill, E.  and Reilly, D.  and Roux, M.  and Shamu, C. E.  and Shang, C. A.  and Steinbeck, C.  and Trefethen, A.  and Williams-Jones, B.  and Wolstencroft, K.  and Xenarios, I.  and Hide, W.},
+   Title={Toward interoperable bioscience data},
+   Journal={Nature Genetics},
+   Year=2012,
+   Volume=44,
+   Number={2},
+   Pages={121--126},
+   doi={10.1038/ng.1054}
 }
 
 @Article{Wilkinson:2016,
[…]
