### 2019-08-20 #1610 @kyleniemeyer
Hi @holtgrewe, I just made some minor fixes to the paper bib file, if you could merge these: https://github.com/bihealth/altamisa/pull/67
