+++ paper.bib
@@ -33,7 +33,7 @@
 @ARTICLE{Gansner:1993,
     author = {Emden R. Gansner and Eleftherios Koutsofios and Stephen C. North and Kiem-Phong Vo},
     title = {A Technique for Drawing Directed Graphs},
-    journal = {{IEEE} Transactions on Software Engineering}
+    journal = {{IEEE} Transactions on Software Engineering},
     year = {1993},
     pages = {214--230},
     doi = {10.1109/32.221135},
