### 2019-08-20 #1610 @kyleniemeyer
Congrats @holtgrewe on your article's publication in JOSS! Thanks to @tleonardi, @dsurujon, and @dmnfarrell for reviewing, and @lpantano for editing.
